requirements.txt
python-telegram-bot==20.0
requests
APScheduler
python-dotenv
nest_asyncio
flask
matplotlib
pandas